+++ retail/BetterBlizzFrames.lua
@@ -61,6 +61,9 @@ local defaultSettings = {
     playerEliteFrameMode = 1,
     hideObjectiveTracker = true,
 
+    rpNames = true,
+    rpNamesFirst = true,
+
     --Target castbar
     playerCastbarIconXPos = 0,
     playerCastbarIconYPos = 0,
@@ -2691,6 +2694,7 @@ end)
 
 local LSM = LibStub("LibSharedMedia-3.0")
 BBF.LSM = LSM
+BBF.allLocales = LSM.LOCALE_BIT_western+LSM.LOCALE_BIT_ruRU+LSM.LOCALE_BIT_zhCN+LSM.LOCALE_BIT_zhTW+LSM.LOCALE_BIT_koKR
 LSM:Register("statusbar", "Blizzard DF", [[Interface\TargetingFrame\UI-TargetingFrame-BarFill]])
 LSM:Register("statusbar", "Blizzard CF", [[Interface\AddOns\BetterBlizzFrames\media\ui-statusbar-cf]])